pyproject.toml
[tool.poetry]
name = "complimentgenbot"
version = "0.1.0"
description = "Compliment generating bot."
authors = ["Constantine Molchanov <moigagoo@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.7"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
pylint = "^2.11.1"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
